requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.1
Adafruit-SSD1306==1.6.1
arrow==0.10.0
astral==1.4
click==6.7
Flask==0.12.2
gpiozero==1.4.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
python-dateutil==2.6.1
pytz==2017.2
rpi-ws281x==3.0.3
RPi.GPIO==0.6.3
six==1.11.0
spidev==3.2
Werkzeug==0.12.2
